+++ lib/src/cache/image_cache_manager_stub.dart
@@ -6,10 +6,11 @@ class ImageCacheManager {
   static final ImageCacheManager instance = ImageCacheManager._();
   ImageCacheManager._();
 
-  ImageLoadStrategy? getStrategy(String url) => throw UnimplementedError();
-  void putStrategy(String url, ImageLoadStrategy strategy) =>
+  ImageLoadStrategy? getStrategy(String key) => throw UnimplementedError();
+  void putStrategy(String key, ImageLoadStrategy strategy) =>
       throw UnimplementedError();
-  Uint8List? getBytes(String url) => throw UnimplementedError();
-  void putBytes(String url, Uint8List bytes) => throw UnimplementedError();
+  void removeStrategy(String key) => throw UnimplementedError();
+  Uint8List? getBytes(String key) => throw UnimplementedError();
+  void putBytes(String key, Uint8List bytes) => throw UnimplementedError();
   void clear() => throw UnimplementedError();
 }